LPS-158949 Rename okta request
