+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/onboarding/context/index.js
@@ -22,7 +22,7 @@ import {
 	getLiferayExperienceCloudEnvironments,
 	getUserAccount,
 } from '../../../common/services/liferay/graphql/queries';
-import {getCurrentSession} from '../../../common/services/okta/rest/sessions';
+import {getCurrentSession} from '../../../common/services/okta/rest/getCurrentSession';
 import {ROLE_TYPES, ROUTE_TYPES} from '../../../common/utils/constants';
 import {getAccountKey} from '../../../common/utils/getAccountKey';
 import {isValidPage} from '../../../common/utils/page.validation';
